requirements.txt
Django==3.1.4
pymongo==3.12.3
djongo==1.3.6
djangorestframework==3.11.1
dnspython==2.2.1
Pillow==8.0.1
jsonfield==3.1.0